icap-v7-master/utility/requirements.txt
Flask==2.0.0
Werkzeug==2.2
flask-cors==3.0.10
python-Levenshtein==0.12.2
gunicorn==20.1.0
python-dotenv==0.17.1
cleanco==2.2
requests==2.27.1
fuzzywuzzy==0.18.0
python-dateutil==2.8.2
pika==1.3.1
redis==4.5.1
pycountry==22.3.5
geotext==0.4.0
price-parser==0.3.4
pycountry==22.3.5
rapidfuzz==3.5.2
spacy==3.5.4
nltk==3.8.1
numpy==1.22.3
unidecode==1.3.2
black==23.7.0
isort==5.12.0
flake8==6.1.0